move _op_translation files
