+++ python/mxnet/onnx/mx2onnx/__init__.py
@@ -19,5 +19,5 @@
 """ONNX Export module"""
 
 from ._export_model import export_model, get_operator_support
-from ._op_translations import _op_translations_opset12
-from ._op_translations import _op_translations_opset13
+from . import _op_translations_opset12
+from . import _op_translations_opset13
